fix numberInput component
