+++ react-app/src/components/InputPages.tsx
@@ -1,0 +1,32 @@
+import { ChangeEvent, useState } from "react";
+
+interface Props {}
+
+export default function NumInput() {
+  const [numOfPages, setNumOfPages] = useState<number>(0);
+
+  const onPagesChange = (e: ChangeEvent<HTMLInputElement>) => {
+    // In general, use Number.isNaN over global isNaN as isNaN will coerce the value to a number first
+    // which likely isn't desired
+    const numOfPages = !Number.isNaN(e.target.valueAsNumber)
+      ? e.target.valueAsNumber
+      : 0;
+    setNumOfPages(numOfPages);
+  };
+
+  console.log(numOfPages);
+  return (
+    <>
+      <input
+        type="number"
+        className="form-control"
+        name="numberOfCredits"
+        id="numberOfCredits"
+        min="1"
+        max="50"
+        value={numOfPages ?? ""}
+        onChange={onPagesChange}
+      />
+    </>
+  );
+}
